Update haveIBeenPwned module
Deprecate existing and replace with new `pwned.js` module.
